+++ libsparse/output_file.h
@@ -30,10 +30,10 @@ struct output_file* output_file_open_fd(int fd, unsigned int block_size, int64_t
 struct output_file* output_file_open_callback(int (*write)(void*, const void*, size_t), void* priv,
                                               unsigned int block_size, int64_t len, int gz,
                                               int sparse, int chunks, int crc);
-int write_data_chunk(struct output_file* out, unsigned int len, void* data);
-int write_fill_chunk(struct output_file* out, unsigned int len, uint32_t fill_val);
-int write_file_chunk(struct output_file* out, unsigned int len, const char* file, int64_t offset);
-int write_fd_chunk(struct output_file* out, unsigned int len, int fd, int64_t offset);
+int write_data_chunk(struct output_file* out, int64_t len, void* data);
+int write_fill_chunk(struct output_file* out, int64_t len, uint32_t fill_val);
+int write_file_chunk(struct output_file* out, int64_t len, const char* file, int64_t offset);
+int write_fd_chunk(struct output_file* out, int64_t len, int fd, int64_t offset);
 int write_skip_chunk(struct output_file* out, int64_t len);
 void output_file_close(struct output_file* out);
